+++ src/components/molecules/myFormCategory/myFormCategory.tsx
@@ -1,164 +1,85 @@
-"use client";
-
-import { Categories } from "@/types/modules";
-import {
-  Button,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TableRow,
-  Typography,
-  Box,
-  CircularProgress
-} from "@mui/material";
-import DeleteForeverIcon from "@mui/icons-material/DeleteForever";
-import AddBoxIcon from "@mui/icons-material/AddBox";
-import { useRouter } from "next/navigation";
-import { Section } from "../section/section";
+import React, { useRef, useState } from "react";
+import { Box, Grid, TextField } from "@mui/material";
+import MyButtonCancel from "../../atoms/button/my-button-cancel";
+import { useForm } from "react-hook-form";
+import MyButtonSave from "@/components/atoms/button/my-button-save";
+import { createCategory } from "@/utils/api";
+import { yupResolver } from "@hookform/resolvers/yup";
+import * as yup from "yup";
 import toast from "react-hot-toast";
-import { MyConfirmModal } from "@/components/molecules/myConfirmModal/myConfirmModal"
-import { useState, useCallback } from "react";
+import { useRouter } from "next/navigation";
 
-interface MyCategoryTableProps {
-  loading?: boolean;
-  categories: Categories[];
-  setCategories: (categories: Categories[]) => void;
-  deleteCategoryCallBack: (id: string) => void;
+const schema = yup
+  .object({
+    name: yup.string().required(),
+  })
+  .required();
+//type FormData = yup.InferType<typeof schema>;
+
+interface MyFormCategoryProps {
+  name?: string | null;
 }
 
-export default function MyCategoryTable(props: MyCategoryTableProps) {
-  const columns = [{ field: "name", headerName: "Nom" }];
-
+export default function MyFormCategory(props: MyFormCategoryProps) {
+  const {
+    handleSubmit,
+    register,
+    formState: { errors, isValid },
+  } = useForm<MyFormCategoryProps>({ resolver: yupResolver(schema) });
+  const [message, setMessage] = useState("");
+  const categoryNameRef = useRef(null);
   const router = useRouter();
 
-  const [show, setShow] = useState(false);
+  function onSubmit(data: MyFormCategoryProps) {
+    createCategory(data).then((result) => {
+      setMessage("")
+      toast.success('Catégorie ajoutée!')
+      if(categoryNameRef && categoryNameRef.current){
+        categoryNameRef.current.value = ""
+        router.push('/category')
+      }
 
-  const [selectedData, setSelectedData] = useState<Categories | null>();
-  
-  const handleDelete = useCallback(() => {
-    if(selectedData){
-
-      props.deleteCategoryCallBack(selectedData._id)
-
-      props.setCategories( 
-        props.categories.filter(
-          (cat) => {
-            
-              return cat._id != selectedData._id
-            }
-          
-        )
-      );
-
-      toast.success("Catégorie supprimée");
-   }
-
-    setShow(false);
-  }, [props, selectedData])
-
-  const handleCancel = useCallback(() => {
-    setShow(false);
-    setSelectedData(null)
-  }, [])
-
+    })
+    .catch((error) => {
+      console.error(error);
+      setMessage("Error");
+    });
+      
+  }
 
   return (
-    <Section>
-      <TableContainer style={{ marginBottom: '100px' }}>
-        <Table>
-          <TableHead>
-            <TableRow>
-              {columns.map((column) => (
-                <TableCell key={column.field} align="left">
-                  <Typography
-                    sx={{ fontSize: "1.75em", color: "primary.main" }}
-                  >
-                    {column.headerName}
-                  </Typography>
-                </TableCell>
-              ))}
-              <TableCell sx={{ paddingX: "0px" }} align="center">
-                <Button
-                  variant="contained"
-                  color="primary"
-                  onClick={() => router.push("/category/new")}
-                >
-                  <Typography
-                    sx={{
-                      display: "flex",
-                      alignContent: "center",
-                      fontSize: "1em",
-                    }}
-                  >
-                    Ajouter une catégorie
-                    <AddBoxIcon
-                      sx={{
-                        borderRadius: "5px",
-                        marginLeft: "5px",
-                      }}
-                    />
-                  </Typography>
-                </Button>
-              </TableCell>
-            </TableRow>
-          </TableHead>
+    <Box>
+      {message && <p>{message}</p>}
+      <form onSubmit={handleSubmit(onSubmit)}>
+        <Grid container rowGap={3} columnGap={2}>
+          <Grid item xs={12}>
+            <TextField
+              id="name"
+              label="Category Name"
+              variant="outlined"
+              fullWidth
 
-          <TableBody>
-            {props.categories && props.categories.length > 0 ? (
-              props.categories.map((category: Categories) => (
-                <TableRow key={category._id}>
-                  <TableCell align="left">
-                    <Typography sx={{ fontSize: "1.25em" }}>
-                      {category.name}
-                    </Typography>
-                  </TableCell>
-                  <TableCell align="center">
-                    <Button
-                      sx={{ color: "black" }}
-                      onClick={() => {
-                        setShow(true)
-                        setSelectedData(category)
-                      }}
-                    >
-                      <DeleteForeverIcon />
-                    </Button>
-                  </TableCell>
-
-                </TableRow>
-
-              ))
-            ) : (
-
-              <TableRow>
-                <TableCell colSpan={5} align="center">
-                  {
-                    !props.loading &&
-
-                    <Typography sx={{ fontSize: "1.25em" }}>
-                      Aucune catégorie
-                    </Typography>
-
-                  }
-                  {
-                    props.loading && props.loading == true &&
-                    <Box sx={{ display: 'flex', justifyContent: 'center' }}>
-                      <CircularProgress />
-                    </Box>
-                  }
-                </TableCell>
-
-              </TableRow>
-            )}
-          </TableBody>
-        </Table>
-
-
-      </TableContainer>
-
-      <MyConfirmModal show={show} setShow={setShow} title="Attention !" message="Êtes-vous certain(e) de vouloir supprimer cette catégorie?" onCancel={handleCancel} onDelete={handleDelete} />
-
-    </Section>
+              {...register("name")}
+              inputProps={{ ref: categoryNameRef }}
+              error={!!errors.name}
+              helperText={errors.name?.message}
+              required
+            />
+          </Grid>
+          <Grid
+            item
+            xs={12}
+            sx={{
+              display: "flex",
+              justifyContent: "space-between",
+              gap: 2,
+            }}
+          >
+            <MyButtonCancel />
+            <MyButtonSave />
+          </Grid>
+        </Grid>
+      </form>
+    </Box>
   );
 }
